Refactor unit tests so that they can be used as mixins
DbTestCase and HttpTestCase and their utility functions were extracted
and distributed to proper files. Documentation is about to be written on
how to use them.
